feat : 부서별 직원 수 조회 기능 추가
CustomEmployeeRepository 및 DepartmentService에 조회 메서드 추가

+++ src/main/java/team7/hrbank/domain/Department/dto/DepartmentListResponse.java
@@ -1,18 +1,16 @@
 package team7.hrbank.domain.Department.dto;
 
-import team7.hrbank.domain.Department.Department;
-
 import java.util.List;
 
 public record DepartmentListResponse(
     List<DepartmentResponse> content,
     String nextCursor,
-    Integer nextIdAfter,
+    Long nextIdAfter,
     Integer size,
     Long totalElements,
     boolean hasNext
 ) {
-    public DepartmentListResponse(List<DepartmentResponse> content, String nextCursor, Integer nextIdAfter, Integer size, Long totalElements, boolean hasNext) {
+    public DepartmentListResponse(List<DepartmentResponse> content, String nextCursor, Long nextIdAfter, Integer size, Long totalElements, boolean hasNext) {
         this.content = content;
         this.nextCursor = nextCursor;
         this.nextIdAfter = nextIdAfter;
